Add files via upload
This is an asset maker tool created for my capstone project. it creates a folder structure with a common UE4 naming convention that is created for each new asset created. the path is hardcoded for what my team used so to be useable to others it would need to be re named.